flowgrind: fix blockcount output on 64bit
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/projects/flowgrind/trunk@6504 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/common.h
@@ -156,10 +156,10 @@ struct _report
 
 	unsigned long long bytes_read;
 	unsigned long long bytes_written;
-	unsigned long request_blocks_read;
-	unsigned long request_blocks_written;
-	unsigned long response_blocks_read;
-	unsigned long response_blocks_written;
+	unsigned int request_blocks_read;
+	unsigned int request_blocks_written;
+	unsigned int response_blocks_read;
+	unsigned int response_blocks_written;
 
 	double rtt_min, rtt_max, rtt_sum;
 	double iat_min, iat_max, iat_sum;
